Module google_pagespeedonline2::api

Structs

PagespeedApiFormatStringV2
There is no detailed description.

PagespeedApiFormatStringV2Args
List of arguments for the format string.

PagespeedApiFormatStringV2ArgsRects
The screen rectangles being referred to, with dimensions measured in CSS pixels. This is only ever used for SNAPSHOT_RECT arguments. If this is absent for a SNAPSHOT_RECT argument, it means that that argument refers to the entire snapshot.

PagespeedApiFormatStringV2ArgsSecondaryRects
Secondary screen rectangles being referred to, with dimensions measured in CSS pixels. This is only ever used for SNAPSHOT_RECT arguments.

PagespeedApiImageV2
There is no detailed description.

PagespeedApiImageV2PageRect
The region of the page that is captured by this image, with dimensions measured in CSS pixels.

PagespeedapiMethods
A builder providing access to all methods supported on pagespeedapi resources. It is not used directly, but through the Pagespeedonline hub.

PagespeedapiRunpagespeedCall
Runs PageSpeed analysis on the page at the specified URL, and returns PageSpeed scores, a list of suggestions to make that page faster, and other information.

Pagespeedonline
Central instance to access all Pagespeedonline related resource activities

Result
There is no detailed description.

ResultFormattedResults
Localized PageSpeed results. Contains a ruleResults entry for each PageSpeed rule instantiated and run by the server.

ResultFormattedResultsRuleResults
The enum-like identifier for this rule. For instance "EnableKeepAlive" or "AvoidCssImport". Not localized.

ResultFormattedResultsRuleResultsUrlBlocks
List of blocks of URLs. Each block may contain a heading and a list of URLs. Each URL may optionally include additional details.

ResultFormattedResultsRuleResultsUrlBlocksUrls
List of entries that provide information about URLs in the url block. Optional.

ResultPageStats
Summary statistics for the page, such as number of JavaScript bytes, number of HTML bytes, etc.

ResultRuleGroups
The name of this rule group: one of "SPEED" or "USABILITY".

ResultVersion
The version of PageSpeed used to generate these results.