Struct google-pagespeedonline2::PagespeedApiFormatStringV2ArgsSecondaryRects

pub struct PagespeedApiFormatStringV2ArgsSecondaryRects {
    pub width: i32,
    pub top: i32,
    pub height: i32,
    pub left: i32,
}

Secondary screen rectangles being referred to, with dimensions measured in CSS pixels. This is only ever used for SNAPSHOT_RECT arguments.

This type is not used in any activity, and only used as part of another schema.

Fields

width

The width of the rect.

top

The top coordinate of the rect, in page coordinates.

height

The height of the rect.

left

The left coordinate of the rect, in page coordinates.