Struct google_pagespeedonline2::ResultFormattedResultsRuleResults

pub struct ResultFormattedResultsRuleResults {
    pub localized_rule_name: Option<String>,
    pub url_blocks: Option<Vec<ResultFormattedResultsRuleResultsUrlBlocks>>,
    pub groups: Option<Vec<String>>,
    pub rule_impact: Option<f64>,
    pub summary: Option<PagespeedApiFormatStringV2>,
}

The enum-like identifier for this rule. For instance "EnableKeepAlive" or "AvoidCssImport". Not localized.

This type is not used in any activity, and only used as part of another schema.

Fields

localized_rule_name

Localized name of the rule, intended for presentation to a user.

url_blocks

List of blocks of URLs. Each block may contain a heading and a list of URLs. Each URL may optionally include additional details.

groups

List of rule groups that this rule belongs to. Each entry in the list is one of "SPEED" or "USABILITY".

rule_impact

The impact (unbounded floating point value) that implementing the suggestions for this rule would have on making the page faster. Impact is comparable between rules to determine which rule's suggestions would have a higher or lower impact on making a page faster. For instance, if enabling compression would save 1MB, while optimizing images would save 500kB, the enable compression rule would have 2x the impact of the image optimization rule, all other things being equal.

summary

A brief summary description for the rule, indicating at a high level what should be done to follow the rule and what benefit can be gained by doing so.