Struct google-pagespeedonline2::ResultType

pub struct ResultType {
    pub kind: String,
    pub response_code: i32,
    pub screenshot: PagespeedApiImageV2,
    pub title: String,
    pub rule_groups: HashMap<String, ResultRuleGroups>,
    pub version: ResultVersion,
    pub formatted_results: ResultFormattedResults,
    pub invalid_rules: Vec<String>,
    pub page_stats: ResultPageStats,
    pub id: String,
}

There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

runpagespeed pagespeedapi (response)

Fields

kind

Kind of result.

response_code

Response code for the document. 200 indicates a normal page load. 4xx/5xx indicates an error.

screenshot

Base64-encoded screenshot of the page that was analyzed.

title

Title of the page, as displayed in the browser's title bar.

rule_groups

A map with one entry for each rule group in these results.

version

The version of PageSpeed used to generate these results.

formatted_results

Localized PageSpeed results. Contains a ruleResults entry for each PageSpeed rule instantiated and run by the server.

invalid_rules

List of rules that were specified in the request, but which the server did not know how to instantiate.

page_stats

Summary statistics for the page, such as number of JavaScript bytes, number of HTML bytes, etc.

id

Canonicalized and final URL for the document, after following page redirects (if any).