Struct google_pagespeedonline2::ResultPageStats

pub struct ResultPageStats {
    pub flash_response_bytes: Option<String>,
    pub total_request_bytes: Option<String>,
    pub number_css_resources: Option<i32>,
    pub text_response_bytes: Option<String>,
    pub number_resources: Option<i32>,
    pub other_response_bytes: Option<String>,
    pub image_response_bytes: Option<String>,
    pub number_hosts: Option<i32>,
    pub javascript_response_bytes: Option<String>,
    pub html_response_bytes: Option<String>,
    pub css_response_bytes: Option<String>,
    pub number_js_resources: Option<i32>,
    pub number_static_resources: Option<i32>,
}

Summary statistics for the page, such as number of JavaScript bytes, number of HTML bytes, etc.

This type is not used in any activity, and only used as part of another schema.

Fields

flash_response_bytes

Number of response bytes for flash resources on the page.

total_request_bytes

Total size of all request bytes sent by the page.

number_css_resources

Number of CSS resources referenced by the page.

text_response_bytes

Number of uncompressed response bytes for text resources not covered by other statistics (i.e non-HTML, non-script, non-CSS resources) on the page.

number_resources

Number of HTTP resources loaded by the page.

other_response_bytes

Number of response bytes for other resources on the page.

image_response_bytes

Number of response bytes for image resources on the page.

number_hosts

Number of unique hosts referenced by the page.

javascript_response_bytes

Number of uncompressed response bytes for JS resources on the page.

html_response_bytes

Number of uncompressed response bytes for the main HTML document and all iframes on the page.

css_response_bytes

Number of uncompressed response bytes for CSS resources on the page.

number_js_resources

Number of JavaScript resources referenced by the page.

number_static_resources

Number of static (i.e. cacheable) resources on the page.